add username to reviews table and require for update and delete Co-authored-by: Morgan Waites <morgan.waites@gmail.com> Co-authored-by: Christopher Neal <christopher.m.neal@gmail.com>

+++ app/controllers/reviews_controller.rb
@@ -1,9 +1,11 @@
 class ReviewsController < ApplicationController
+   
   def index
     @reviews = Review.all
     ratings = params[:ratings]
     most = params[:most_reviews]
     city = params[:city]
+    random = params[:random]
     country = params[:country]
     if city
       @reviews = Review.city_search(city)
@@ -13,6 +15,8 @@ class ReviewsController < ApplicationController
       @reviews = Review.best_rating
     elsif most
       @reviews = Review.most_reviews
+    elsif random
+      @reviews = Review.random
     end
     json_response(@reviews)
   end
@@ -28,24 +32,34 @@ class ReviewsController < ApplicationController
   end
 
   def update
+    user_name = params[:user_name]
     @review = Review.find(params[:id])
-    if @review.update!(review_params)
-      render status: 200, json: {
-      message: "This review has been updated successfully."
-      }
+    if user_name == @review.user_name
+      if @review.update!(review_params)
+        render status: 200, json: {
+        message: "This review has been updated successfully."
+        }
+      end
+    else
+      render json: { error: "You are unauthorized" }
     end
   end
 
   def destroy
+    user_name = params[:user_name]
     @review = Review.find(params[:id])
-    if @review.destroy
-      render status: 200, json: {
-      message: "This review has been deleted successfully."
-      }
+    if user_name == @review.user_name
+      if @review.destroy
+        render status: 200, json: {
+        message: "This review has been deleted successfully."
+        }
+      end
+    else
+      render json: { error: "You are not the autherized user" }
     end
   end
 
   def review_params
-    params.permit(:content, :author, :rating, :landmark, :city, :country)
+    params.permit(:content, :author, :rating, :landmark, :city, :country, :user_name)
   end
 end
